database_postgres_local/requirements.txt
# Dependências do projeto
pandas>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0  # Para carregar variáveis de ambiente
